Fixed the ::getJavadoc function Added tests, and renamed the utils file
