+++ test/test-utils.js
@@ -1,0 +1,30 @@
+/* global describe, beforeEach, it*/
+
+const assert = require('yeoman-assert');
+const utils = require('../generators/utils');
+
+describe('JHipster Utils', () => {
+    describe('::getJavadoc', () => {
+        describe('when passing a negative or nil increment', () => {
+            it('returns the comment with no increment', () => {
+                assert.textEqual(utils.getJavadoc('whatever', -42), '/**\n * whatever\n */');
+                assert.textEqual(utils.getJavadoc('whatever', 0), '/**\n * whatever\n */');
+            });
+        });
+        describe('when passing a positive increment', () => {
+            it('returns the comment with the increment', () => {
+                assert.textEqual(utils.getJavadoc('whatever', 1), ' /**\n  * whatever\n  */');
+            });
+        });
+        describe('when passing a nil comment', () => {
+            it('inserts an empty comment instead of failing', () => {
+                assert.textEqual(utils.getJavadoc(null, 1), ' /**\n  * \n  */');
+            });
+        });
+        describe('when passing a comment containing double quotes', () => {
+            it('escapes the quotes', () => {
+                assert.textEqual(utils.getJavadoc('Comment="KO"', 1), ' /**\n  * Comment=\\"KO\\"\n  */');
+            });
+        });
+    });
+});
